requirements.txt
pandas==2.2.2
matplotlib==3.9.0
mysql-connector-python==8.4.0
jupyter==1.0.0
